chore(acir): organise opcodes definitions
